Fixed inheritance problem in CUTErTestProblem.
o  CUTErTestProblem.__init__() was not calling
   OptimizationTestProblem.__init__() properly.
o  The typical name of the CUTEr database if CLASSF.DB. Fixed in
   cuterfactory.py.
o  Added a description member to TestProblem class.
o  Added a remove_problem() method to ProblemSet class.
o  Added a pop_problem() method to ProblemSet class.
o  Changed input of ModelStructure class to a problem is unconstrained
   by default.